+++ openfisca_aotearoa/variables/acts/income_tax/interpretation.py
@@ -1,7 +1,9 @@
 """TODO: Add missing doctring."""
 
 # Import from openfisca-core the common python objects used to code the legislation in OpenFisca
-from openfisca_core import holders, periods, variables
+
+from openfisca_core import holders
+from openfisca_core import periods, variables
 
 from openfisca_aotearoa import entities