+++ openfisca_aotearoa/variables/acts/social_security/interpretation/child.py
@@ -2,7 +2,8 @@
 
 import numpy
 
-from openfisca_core import holders, periods, variables
+from openfisca_core import holders
+from openfisca_core import periods, variables
 
 from openfisca_aotearoa import entities