+++ openfisca_aotearoa/variables/acts/social_security/interpretation/relationship.py
@@ -1,5 +1,6 @@
 """This module provides eligibility and amount for Jobseeker Support."""
 
+from openfisca_core import holders
 from openfisca_core import periods, variables
 
 from openfisca_aotearoa import entities
@@ -18,3 +19,12 @@ class social_security__parent(variables.Variable):
     entity = entities.Person
     definition_period = periods.MONTH
     label = "Is a parent?"
+
+
+class social_security__principal_caregiver(variables.Variable):
+    value_type = bool
+    entity = entities.Person
+    definition_period = periods.MONTH
+    label = "Principal Caregiver in relation to a dependent child"
+    set_input = holders.set_input_dispatch_by_period
+    reference = "https://legislation.govt.nz/act/public/1964/0136/latest/whole.html#DLM6784656"