+++ openfisca_aotearoa/variables/acts/social_security/schedule_4/part1.py
@@ -1,11 +1,11 @@
 """This module provides eligibility and amount for Jobseeker Support."""
 
+import numpy
+
 # We import the required OpenFisca modules needed to define a formula.
 #
 # For more information on OpenFisca's available modules:
 # https://openfisca.org/doc/openfisca-python-api/index.html
-import numpy
-
 from openfisca_core import periods, variables
 
 # We import the required `entities` corresponding to our formulas.